First alpha release version of GGGain
this is the first version, I think better psychoaccoustic modeling might be possible or other considerations. I'm not sure if the concept has any practical purpose yet but I thought it was an interesting concept to havea  tool to see how hearing loss affects how pepople hear music differentyl over time, this is not likely perfect science but a proof of concept.